[FIX/#71] SetDateView 리펙토링 처리
